requirements.txt
asgiref==3.5.2
black==22.6.0
boto3==1.24.56
botocore==1.27.56
click==8.1.3
crispy-bootstrap5==0.6
Django==4.1
django-appconf==1.0.5
django-avatar==7.0.0
django-crispy-forms==1.14.0
django-leaflet==0.28.3
django-phonenumber-field==7.0.0
django-storages==1.13.1
djangorestframework==3.13.1
docutils==0.19
Faker==14.1.0
gunicorn==20.1.0
jmespath==1.0.1
mypy-extensions==0.4.3
pathspec==0.10.0
phonenumberslite==8.12.55
Pillow==9.2.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.2.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.11
